Update next injector
