--- scripts/events/index.js
@@ -1,7 +1,0 @@
-/* global hexo */
-
-'use strict';
-
-hexo.extend.filter.register('before_generate', () => {
-  require('./lib/injects')(hexo);
-});
